Feat: Todo, Post 전체 조회
전체 조회 기능 추가

+++ lib/src/data/models/user.dart
@@ -45,7 +45,7 @@ class User {
 
   factory User.fromJson(Map<String, dynamic> json) {
     return User(
-      id: json['id'],
+      id: json['id'] as int,
       name: json['name'],
       username: json['username'],
       email: json['email'],
